Move CLI utilities into internal/util/cli from internal/util to balance tests
